Added the new MaskBathymetry class
Added a new implementation of the Bathymetry class, the
MaskBathymetry class.

The MaskBathymetry generates a bathymetry map by taking the depth
of the bottom face of the last cell of a mask. The quality of the
bathymetry chart, therefore, depends on the quality of the
meshmask and on its level of refinement