caching result of the is_regular method for a Mask
The is_regular method is computationally quite expensive; some
functions (for example `convert_lon_lat_to_indices`) call this
method on each invocation (and so every time we check if the
grid is regular or not).

Now this check happens one, the result is cached and the method
is_regular simply returns the value of the cache.